// mechanics-package/src/dragAndDrop/items/equipment/types.ts
/**
 * Общий тип со всеми видами экипировки
 */
export type myEquipmentType = "helmet" | "gloves" | "bracers" | "shoulderPads" | "oneHandWeapon"
    | "twoHandWeapon" | "offHand" | "chestArmor" | "belt" | "footArmor"
    | "boots";

/**
 * Общий тип со всеми ячейками экипировки
 */
export type myEquipmentContainer = "helmet" | "gloves" | "bracers" | "shoulderPads" | "leftHand"
    | "rightHand" | "chestArmor" | "belt" | "footArmor" | "boots";